move NameTableMirroring enum to cartridge and change return type of getNameTableMirroring
